+++ src/PrestControl.Views/ClienteFormView.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows.Controls;
+
+namespace PrestControl.Views;
+
+public partial class ClienteFormView : UserControl
+{
+    public ClienteFormView() => InitializeComponent();
+}